+++ packages/dock-blockchain-modules/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @docknetwork/dock-blockchain-modules
 
+## 0.22.0
+
+### Minor Changes
+
+- Update id mapping to reflect latest `testnet` state, misc tweaks
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.33.0
+
 ## 0.21.0
 
 ### Minor Changes